Moves everything to one template
